Prep: move apply_headers to remote_provider_reapi
